requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==1.11.11
djangorestframework==3.9.0
kombu==4.6.3
mysqlclient==1.3.13
numpy==1.17.0
pandas==0.25.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
vine==1.3.0
